requirements.txt
instaloader==4.10.3
requests==2.31.0
python-dotenv==1.0.0
PyJWT==2.8.0
cryptography==41.0.7
